updated code structure
