+++ .readthedocs.yaml
@@ -1,9 +1,9 @@
 version: 2
 
 build:
-  os: "ubuntu-20.04"
+  os: "ubuntu-22.04"
   tools:
-    python: "3.9"
+    python: "3.11"
 
 sphinx:
   configuration: docs/source/conf.py
@@ -13,4 +13,4 @@ python:
     - method: pip
       path: .
       extra_requirements:
-        - doc
+        - docs
