+++ docs/source/tutorials/index.md
@@ -5,24 +5,24 @@ To faciliate the use of our package, we provide a few tutorials to demonstrate t
 - [Using `DifferentiableCollisions.jl` in Julia and Python](julia.md)
 - [Controller implementation](controller.md)
 
-To run the three simulation examples, first following the installation process, then `cd` into the folder `/path/to/DifferentiableOptimizationCBF/DifferentiableOptimizationCBF` (**change `/path/to/` to the path you cloned `DifferentiableOptimizationCBF` into**) and run the following commands in the terminal:
+To run the three simulation examples, first follow the installation process, then run the following commands from the repository root:
 
 ```bash
 # For the unicycle example, run the following command
-python3 unicycle_exp.py
+uv run diffoptcbf-unicycle
 
 # For the three-blocks example, run the following command
-python3 three_blocks_exp.py
+uv run diffoptcbf-three-blocks
 
 # For the two-walls example, run the following command
-python3 two_walls_exp.py
+uv run diffoptcbf-two-walls
 ```
 
 Note that because of the JIT compilation of Julia, it takes a while to start the simulation.
 
 ## FAQ
 
-- If you see the following error message `Your Python interpreter /usr/bin/python3 is statically linked to libpython. Currently, PyJulia does not fully support such Python interpreter.` Try changing `python3` to `python-jl`.
+- If Julia/Python interop reports a statically linked Python interpreter, make sure you are running inside the `uv` environment created by `uv sync`.
 - If you see `Cannot connect to X server`, on the host machine try running `xhost +`.
 
 ```{toctree}
