+++ src/DifferentiableOptimizationCBF/dc_utils/create_arm_capsule.jl
