+++ src/DifferentiableOptimizationCBF/dc_utils/create_arm_ellipsoid.jl
