+++ src/DifferentiableOptimizationCBF/dc_utils/get_cbf_link_polygon_obstacle.jl
