+++ src/DifferentiableOptimizationCBF/dc_utils/get_cbf_three_blocks.jl
