+++ src/DifferentiableOptimizationCBF/dc_utils/get_cbf_two_walls.jl
