+++ src/DifferentiableOptimizationCBF/dc_utils/three_blocks_exp_setup.jl
