+++ src/DifferentiableOptimizationCBF/robots/fr3_hand_collision.urdf
