+++ src/DifferentiableOptimizationCBF/robots/fr3_link3_ellipsoid.urdf
