+++ src/DifferentiableOptimizationCBF/robots/fr3_link4_capsule.urdf
