+++ src/DifferentiableOptimizationCBF/robots/fr3_link4_ellipsoid.urdf
