+++ src/DifferentiableOptimizationCBF/robots/fr3_link5_1_capsule.urdf
