+++ src/DifferentiableOptimizationCBF/robots/fr3_link5_1_ellipsoid.urdf
