+++ src/DifferentiableOptimizationCBF/robots/fr3_link5_2_ellipsoid.urdf
