+++ src/DifferentiableOptimizationCBF/robots/fr3_link6_capsule.urdf
