+++ src/DifferentiableOptimizationCBF/robots/fr3_link6_ellipsoid.urdf
