+++ src/DifferentiableOptimizationCBF/robots/fr3_link7_ellipsoid.urdf
