+++ src/DifferentiableOptimizationCBF/robots/meshes/fr3_link3_ellipsoid.obj
