+++ src/DifferentiableOptimizationCBF/robots/meshes/fr3_link4_ellipsoid.obj
