+++ src/DifferentiableOptimizationCBF/robots/meshes/fr3_link5_1_ellipsoid.obj
