+++ src/DifferentiableOptimizationCBF/robots/meshes/fr3_link5_2_ellipsoid.obj
