+++ src/DifferentiableOptimizationCBF/robots/meshes/fr3_link6_ellipsoid.obj
