+++ src/DifferentiableOptimizationCBF/robots/meshes/fr3_link7_ellipsoid.obj
