+++ src/DifferentiableOptimizationCBF/robots/square_obstacle.urdf
