+++ src/DifferentiableOptimizationCBF/robots/wall_back_lower.urdf
