+++ src/DifferentiableOptimizationCBF/robots/wall_back_upper.urdf
