+++ src/DifferentiableOptimizationCBF/robots/wall_front_lower.urdf
