+++ src/DifferentiableOptimizationCBF/robots/wall_front_upper.urdf
