+++ src/DifferentiableOptimizationCBF/three_blocks_controller.py
