+++ src/DifferentiableOptimizationCBF/two_walls_exp.py
@@ -4,18 +4,20 @@ import numpy as np
 
 
 def main():
+    from DifferentiableOptimizationCBF.envs.two_walls_env import TwoWallsEnv
+    from DifferentiableOptimizationCBF.two_walls_controller import TwoWallsController
+
     # create environment
-    env = ThreeBlocksEnv(
+    env = TwoWallsEnv(
         render_mode="human",
         record_path=None,
-        crude_type="ellipsoid",
     )
 
-    controller = ThreeBlocksController()
+    controller = TwoWallsController()
 
     # reset environment
     info = env.reset(
-        cameraDistance=2.0, cameraYaw=-1e-3, cameraPitch=-1e-3, lookat=[0.45, 0.0, 0.55]
+        cameraDistance=2.0, cameraYaw=-1e-3, cameraPitch=-1e-3, lookat=[0.70, 0.0, 0.55]
     )
 
     # initialize clock
@@ -47,7 +49,11 @@ def main():
             history.append(_info)
 
         # compute torque command
-        τ = 6.0 * (dq_target[:, np.newaxis] - dq[:, np.newaxis]) + G - 0.1 * dq[:, np.newaxis]
+        τ = (
+            6.0 * (dq_target[:, np.newaxis] - dq[:, np.newaxis])
+            + G
+            - 0.1 * dq[:, np.newaxis]
+        )
         torques.append(τ)
 
         if i >= 1:
@@ -58,9 +64,4 @@ def main():
 
 
 if __name__ == "__main__":
-    from DifferentiableOptimizationCBF.envs.three_blocks_env import ThreeBlocksEnv
-    from DifferentiableOptimizationCBF.three_blocks_controller import (
-        ThreeBlocksController,
-    )
-
     main()
